requirements.txt
google-adk
google-cloud-aiplatform
google-cloud-storage
python-dotenv